Added parameters for tests (experiment-5)
